Move resolveLinkDestination out of the component file
React Doctor flagged the non-component export in LinkDestinationBar.tsx
(only-export-components). The URL-resolution rules now live in
shell/linkDestination.ts; the component file exports only components.

Generated-By: PostHog Desktop
Task-Id: a7502615-3522-4d04-93ed-3bb33534c1b2
